requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyRAPL>=0.2.3
psutil>=5.9.0
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
